Finished MotoHub

+++ MotoHub/Services/MotorcycleService.cs
@@ -1,7 +1,10 @@
 ﻿using AutoMapper;
+using MotoHub.CrossCutting;
 using MotoHub.DTOs;
+using MotoHub.Entities;
 using MotoHub.Models;
 using MotoHub.Repositories;
+using MotoHub.Services.RabbitMQ;
 
 namespace MotoHub.Services
 {
@@ -9,11 +12,19 @@ namespace MotoHub.Services
     {
         private readonly IMotorcycleRepository _repository;
         private readonly IMapper _mapper;
+        private readonly IMessagingPublisherService _messagingPublisherService;
+        private readonly IRentalOperationService _rentalOperationService;
 
-        public MotorcycleService(IMotorcycleRepository repository, IMapper mapper)
+        public MotorcycleService(
+            IMotorcycleRepository repository,
+            IMapper mapper,
+            IMessagingPublisherService messagingPublisherService,
+            IRentalOperationService rentalOperationService)
         {
             _repository = repository;
             _mapper = mapper;
+            _messagingPublisherService = messagingPublisherService;
+            _rentalOperationService = rentalOperationService;
         }
 
         public IEnumerable<MotorcycleDTO> GetAllMotorcycles()
@@ -22,9 +33,9 @@ namespace MotoHub.Services
             return _mapper.Map<IEnumerable<MotorcycleDTO>>(motorcycles);
         }
 
-        public MotorcycleDTO GetMotorcycleByLicensePlate(string licensePlate)
+        public async Task<MotorcycleDTO> GetMotorcycleByLicensePlateAsync(string licensePlate)
         {
-            var motorcycle = _repository.GetByLicensePlate(licensePlate);
+            var motorcycle = await _repository.GetByLicensePlateAsync(licensePlate);
             return _mapper.Map<MotorcycleDTO>(motorcycle);
         }
 
@@ -34,40 +45,46 @@ namespace MotoHub.Services
             _repository.Add(motorcycle);
         }
 
-        public void UpdateMotorcycle(string licensePlate, MotorcycleDTO motorcycleDto)
+        public async Task UpdateMotorcycleAsync(string licensePlate, string newLicencePlate)
         {
-            var existingMotorcycle = _repository.GetByLicensePlate(licensePlate);
+            var existingMotorcycle = await _repository.GetByLicensePlateAsync(licensePlate);
             if (existingMotorcycle == null)
             {
-                // Lidar com a situação em que a motocicleta não existe
                 return;
             }
 
-            existingMotorcycle.Year = motorcycleDto.Year;
-            existingMotorcycle.Model = motorcycleDto.Model;
-            existingMotorcycle.LicensePlate = motorcycleDto.LicensePlate;
+            existingMotorcycle.LicensePlate = newLicencePlate;
 
             _repository.Update(existingMotorcycle);
+
+            LicencePlateRabbitMQEntity licencePlateRabbitMQEntity = new LicencePlateRabbitMQEntity()
+            {
+                newLicencePlate = newLicencePlate,
+                oldLicencePlate = licensePlate,
+            };
+
+            _messagingPublisherService.PublishLicenceUpdate(licencePlateRabbitMQEntity);
         }
 
-        public void DeleteMotorcycle(string licensePlate)
+        public async Task<OperationResult> DeleteMotorcycle(string licensePlate)
         {
-            var existingMotorcycle = _repository.GetByLicensePlate(licensePlate);
+            var existingMotorcycle = await _repository.GetByLicensePlateAsync(licensePlate);
             if (existingMotorcycle == null)
+                return OperationResult.Fail($"Motorcycle with plate {licensePlate} not found.");
+
+            var isRented = await _rentalOperationService.GetRentalsByMotorcycleLicencePlateAsync(licensePlate);
+            if (isRented)
+                return OperationResult.Fail("Motorcycle is currently rented and cannot be deleted.");
+
+            try
             {
-                // Lidar com a situação em que a motocicleta não existe
-                return;
+                _repository.Delete(existingMotorcycle.Id);
+                return OperationResult.Ok("Motorcycle successfully deleted.");
             }
-
-            // Lidar com a situação em que a motocicleta tem locações ativas no microserviço de locação
-            // Remova o comentário quando o microserviço de locação estiver pronto para lidar com isso
-            /*if (_repository.HasActiveRentals(existingMotorcycle.Id))
+            catch (Exception ex)
             {
-                // Lidar com a situação em que a motocicleta tem locações ativas
-                return;
-            }*/
-
-            _repository.Delete(existingMotorcycle.Id);
+                return OperationResult.Fail("Failed to delete the motorcycle due to an unexpected error." + ex.Message);
+            }
         }
